feat(vibrant-core): add text systemProp
