+++ packages/vibrant-components/src/lib/Text/TextProps.ts
@@ -1,9 +1,9 @@
 import type { ReactNode, RefObject } from 'react';
-import type { ResponsiveValue, SizingProps, TypographyProps } from '@vibrant-ui/core';
+import type { ResponsiveValue, SizingProps, TextProps as SystemTextProps } from '@vibrant-ui/core';
 import { withVariation } from '@vibrant-ui/core';
 import type { ColorToken, TypographyKind, TypographyWeight } from '@vibrant-ui/theme';
 
-export type TextProps = Omit<TypographyProps, 'fontWeight' | 'typography'> &
+export type TextProps = SystemTextProps &
   Pick<SizingProps, 'maxWidth'> & {
     as?: 'p' | 'span';
     ref?: RefObject<HTMLElement>;
