+++ packages/vibrant-core/src/index.ts
@@ -14,6 +14,7 @@ export type {
   PseudoClassProps,
   SizingProps,
   SpacingProps,
+  TextProps,
   TypographyProps,
   SystemProps,
   ElevationProps,
