+++ packages/vibrant-core/src/lib/props/text/index.ts
@@ -1,0 +1,2 @@
+export type { TextProps } from './type';
+export { textProps } from './text';
