+++ packages/vibrant-core/src/lib/props/text/type.ts
@@ -1,0 +1,11 @@
+import type { ResponsiveValue } from '../../../types';
+
+export type TextProps = {
+  letterSpacing?: ResponsiveValue<number | string>;
+  textAlign?: ResponsiveValue<'center' | 'end' | 'justify' | 'left' | 'match-parent' | 'right' | 'start'>;
+  textTransform?: ResponsiveValue<'capitalize' | 'full-size-kana' | 'full-width' | 'lowercase' | 'none' | 'uppercase'>;
+  whiteSpace?: ResponsiveValue<'break-spaces' | 'normal' | 'nowrap' | 'pre-line' | 'pre-wrap' | 'pre'>;
+  wordBreak?: ResponsiveValue<'break-all' | 'break-word' | 'keep-all' | 'normal'>;
+  wordWrap?: ResponsiveValue<'break-word' | 'normal'>;
+  lineLimit?: number;
+};
