+++ packages/vibrant-core/src/lib/props/typography/typography.ts
@@ -17,11 +17,6 @@ const fontSizeProp = createSystemProp({
   property: 'fontSize',
 });
 
-const lineHeightProp = createSystemProp({
-  property: 'lineHeight',
-  transform: (value: number | string) => ({ lineHeight: typeof value === 'number' ? `${value}px` : value }),
-});
-
 const fontWeightProp = createSystemProp({
   property: 'fontWeight',
   scale: 'typographyWeight',
@@ -31,53 +26,16 @@ const fontStyleProp = createSystemProp({
   property: 'fontStyle',
 });
 
-const letterSpacingProp = createSystemProp({
-  property: 'letterSpacing',
-});
-
-const textAlignProp = createSystemProp({
-  property: 'textAlign',
-});
-
-const textTransformProp = createSystemProp({
-  property: 'textTransform',
-});
-
-const whiteSpaceProp = createSystemProp({
-  property: 'whiteSpace',
-});
-
-const wordBreakProp = createSystemProp({
-  property: 'wordBreak',
-});
-
-const wordWrapProp = createSystemProp({
-  property: 'wordWrap',
-});
-
-const lineLimitProp = createSystemProp({
-  property: 'lineLimit',
-  transform: (value: number) => ({
-    display: '-webkit-box',
-    overflow: 'hidden',
-    textOverflow: 'ellipsis',
-    WebkitBoxOrient: 'vertical',
-    WebkitLineClamp: value,
-  }),
+const lineHeightProp = createSystemProp({
+  property: 'lineHeight',
+  transform: (value: number | string) => ({ lineHeight: typeof value === 'number' ? `${value}px` : value }),
 });
 
 export const typographyProps = [
   typographyProp,
+  fontFamilyProp,
+  fontSizeProp,
   fontWeightProp,
   fontStyleProp,
-  letterSpacingProp,
-  textAlignProp,
-  textTransformProp,
-  fontFamilyProp,
   lineHeightProp,
-  fontSizeProp,
-  whiteSpaceProp,
-  wordBreakProp,
-  wordWrapProp,
-  lineLimitProp,
 ];
